rename rmatrix_solver to core, fix typo i asymptotics.decoupled
