+++ src/jitr/rmatrix/__init__.py
@@ -1,5 +1,2 @@
 from .rmatrix import Solver
-from .rmatrix_solver import (
-    solution_coeffs_with_inverse,
-    solve_smatrix_with_inverse,
-)
+from . import core
