+++ src/jitr/rmatrix/core.py
