+++ src/jitr/rmatrix/rmatrix.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from ..reactions.system import Channels, Asymptotics
 from ..utils import block
-from .rmatrix_solver import (
+from .core import (
     solution_coeffs_with_inverse,
     solve_smatrix_with_inverse,
 )
@@ -50,7 +50,7 @@ class Solver:
         return block(matrix, (i, j), (N, N))
 
     def free_matrix(
-        self, a: np.float64, l: np.array, energy_ratio: np.ndarray, coupled=True
+        self, a: np.float64, l: np.array, energy_ratio: np.ndarray=None, coupled=True
     ):
         r"""
         precompute free matrix, which only depend on the channel orbital
@@ -67,6 +67,9 @@ class Solver:
                 number of basis elements, othereise returns the full
                 (Nch x Nb, Nch x Nb) matrix
         """
+        if energy_ratio is None:
+            energy_ratio = np.ones(l.size)
+
         free_matrix = self.kernel.free_matrix(a, l, energy_ratio)
 
         if coupled:
